requirements.txt
pytest>=2.6.4
requests>=2.5.1
exifread>=2.0.0
gpxpy>=0.9.8
python-dateutil>=2.4.0